Improved Interface
Added option to back out of payment/deposit adding and return to the
main menu. Currently, this only works for the inputting of dates and
times. After that point, it is assumed that the desired transaction
should be completed. This can later be adjusted to check on all fields,
but is adequate for now.

Completed project highlists section in README.